+++ resources.go
@@ -15,6 +15,7 @@
 package newrelic
 
 import (
+	"strings"
 	"unicode"
 
 	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
@@ -31,14 +32,24 @@ const (
 	mainPkg = "newrelic"
 	// modules:
 	mainMod       = "index"
-	syntheticsMod = "synthetics"
-	insightsMod   = "insights"
-	pluginsMod    = "plugins"
+	syntheticsMod = "Synthetics"
+	insightsMod   = "Insights"
+	pluginsMod    = "Plugins"
 )
 
-// makeMember manufactures a type token for the package and the given module and type.
-func makeMember(mod string, mem string) tokens.ModuleMember {
-	return tokens.ModuleMember(mainPkg + ":" + mod + ":" + mem)
+var namespaceMap = map[string]string{
+	mainPkg: "NewRelic",
+}
+
+// makeMember manufactures a type token for the package and the given module and type.  It automatically
+//// uses the NewRelic package and names the file by simply lower casing the resource's first character.
+// first character.
+func makeMember(moduleTitle string, mem string) tokens.ModuleMember {
+	moduleName := strings.ToLower(moduleTitle)
+	namespaceMap[moduleName] = moduleTitle
+	fn := string(unicode.ToLower(rune(mem[0]))) + mem[1:]
+	token := moduleName + "/" + fn
+	return tokens.ModuleMember(mainPkg + ":" + token + ":" + mem)
 }
 
 // makeType manufactures a type token for the package and the given module and type.
@@ -46,20 +57,14 @@ func makeType(mod string, typ string) tokens.Type {
 	return tokens.Type(makeMember(mod, typ))
 }
 
-// makeDataSource manufactures a standard resource token given a module and resource name.  It
-// automatically uses the main package and names the file by simply lower casing the data source's
-// first character.
+// makeDataSource manufactures a standard resource token given a module and resource name.
 func makeDataSource(mod string, res string) tokens.ModuleMember {
-	fn := string(unicode.ToLower(rune(res[0]))) + res[1:]
-	return makeMember(mod+"/"+fn, res)
+	return makeMember(mod, res)
 }
 
-// makeResource manufactures a standard resource token given a module and resource name.  It
-// automatically uses the main package and names the file by simply lower casing the resource's
-// first character.
+// makeResource manufactures a standard resource token given a module and resource name.
 func makeResource(mod string, res string) tokens.Type {
-	fn := string(unicode.ToLower(rune(res[0]))) + res[1:]
-	return makeType(mod+"/"+fn, res)
+	return makeType(mod, res)
 }
 
 // preConfigureCallback is called before the providerConfigure function of the underlying provider.
@@ -148,6 +153,7 @@ func Provider() tfbridge.ProviderInfo {
 				"Pulumi":                       "1.5.0-*",
 				"System.Collections.Immutable": "1.6.0",
 			},
+			Namespaces: namespaceMap,
 		},
 	}