+++ sdk/dotnet/Insights/Event.cs
@@ -5,7 +5,7 @@ using System.Collections.Immutable;
 using System.Threading.Tasks;
 using Pulumi.Serialization;
 
-namespace Pulumi.Newrelic.Insights
+namespace Pulumi.NewRelic.Insights
 {
     /// <summary>
     /// ## Events